+++ src/core/ton_wallet/mod.rs
@@ -385,7 +385,7 @@ impl TonWallet {
             .await
     }
 
-    pub async fn estimate_fees(&mut self, message: &ton_block::Message) -> Result<u64> {
+    pub async fn estimate_fees(&mut self, message: &ton_block::Message) -> Result<u128> {
         self.contract_subscription.estimate_fees(message).await
     }
 }
